requirements.txt
langchain
langchain-community
langchain-core
langchain-ollama
langchain-huggingface
sentence-transformers
faiss-cpu
streamlit
faker
python-dotenv
watchdog

langchain-groq
